refactor: do not feature gate decl macro
